scamscan/pyproject.toml
[project]
name = "scamscan-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.27.0",
    "greenlet>=3.2.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
